Updated screeningDay controller, service and repository

+++ CinemaApp.Domain/Services/ScreeningDayService/ScreeningDayService.cs
@@ -1,6 +1,7 @@
 ﻿using CinemaApp.DAL.Repositories.ScreeningDayRepository;
 using CinemaApp.Database.Entities.MovieModels;
 using CinemaApp.Domain.DTO.ScreeningDayDTOModels;
+using CinemaApp.Domain.Exceptions;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -17,28 +18,47 @@ namespace CinemaApp.Domain.Services.ScreeningDayService
             _screeningDayRepository = screeningDayRepository;
         }
 
+        public ScreeningDay GetEntityById(int id)
+        {
+            var screeningDay = _screeningDayRepository.GetEntityById(id);
+            if (screeningDay == null)
+                throw new ItemDoesntExistException();
+
+            return screeningDay;
+        }
+        public IEnumerable<ScreeningDay> GetAllScreeningDays()
+        {
+            var screeningDays = _screeningDayRepository.GetAllScreeningDays().ToList();
+            if (screeningDays == null || screeningDays.Count == 0)
+                throw new ListIsEmptyException();
+
+            return screeningDays;
+        }
+
         public void AddScreeningDay(ScreeningDay screeningDay)
         {
-            //check for errors
+            var screeningDayFromDb = _screeningDayRepository.GetAllScreeningDays().FirstOrDefault(s => s.Date == screeningDay.Date);
+            if (screeningDayFromDb != null)
+                throw new ItemAlreadyExistsException();
+
             _screeningDayRepository.AddScreeningDay(screeningDay);
         }
 
         public void DeleteScreeningDayById(int id)
         {
-            //check for errors
+            var screeningDay = _screeningDayRepository.GetEntityById(id);
+            if (screeningDay == null)
+                throw new ItemDoesntExistException();
+
             _screeningDayRepository.DeleteScreeningDayById(id);
         }
 
-        public IEnumerable<ScreeningDay> GetAllScreeningDays()
+        public IEnumerable<ScreeningDayDTO> GetAllScreeningDayDTOs()
         {
-            //check for errors
-            var screeningDays = _screeningDayRepository.GetAllScreeningDays();
-            return screeningDays;
-        }
+            var screeningDays = _screeningDayRepository.GetAllScreeningDays().ToList();
+            if (screeningDays == null || screeningDays.Count == 0)
+                throw new ListIsEmptyException();
 
-        public IEnumerable<ScreeningDayDTO> GetAllScreeningDaysDTO()
-        {
-            var screeningDays = _screeningDayRepository.GetAllScreeningDays();
             var screeningDaysDTO = new List<ScreeningDayDTO>();
             
             foreach(ScreeningDay screeningDay in screeningDays)
@@ -55,30 +75,23 @@ namespace CinemaApp.Domain.Services.ScreeningDayService
                 foreach (Screening screening in screeningDay.Screenings)
                 {
                     screeningDto.Screenings.Add
-                        (
-                            new ScreeningDTO
+                    (
+                        new ScreeningDTO
+                        {
+                            Id = screening.Id,
+                            Hour = screening.Hour,
+                            Movie = new MovieDTO
                             {
-                                Id = screening.Id,
-                                Hour = screening.Hour,
-                                Movie = new MovieDTO
-                                {
-                                    Id = screening.Movie.Id,
-                                    PictureUrl = screening.Movie.PictureUrl,
-                                    Title = screening.Movie.Title
-                                }
+                                Id = screening.Movie.Id,
+                                PictureUrl = screening.Movie.PictureUrl,
+                                Title = screening.Movie.Title
                             }
-                        );;
+                        }
+                    );
                 }
             }
 
             return screeningDaysDTO;
-        }
-
-        public ScreeningDay GetEntityById(int id)
-        {
-            //check for errors
-            var screeningDay = _screeningDayRepository.GetEntityById(id);
-            return screeningDay;
         }
     }
 }
